Inline Retry-After header value parsing
as it is only used in one place and doesn't need to be exported
