+++ src/http-api/utils.ts
@@ -131,26 +131,6 @@ function getResponseContentType(headers: Headers): ParsedMediaType | null {
 }
 
 /**
- * Parse a Retry-After header value and convert it into a relative delay in milliseconds.
- * @see {@link https://www.rfc-editor.org/rfc/rfc9110#section-10.2.3-2}
- * @throws Error if the provided value is an invalid Retry-After header value
- */
-export function parseRetryAfterMs(retryAfter: string): number {
-    if (/^\d+$/.test(retryAfter)) {
-        const ms = Number.parseInt(retryAfter) * 1000;
-        if (!Number.isFinite(ms)) {
-            throw new Error("integer value is too large");
-        }
-        return ms;
-    }
-    const date = new Date(retryAfter);
-    if (date.toUTCString() !== retryAfter) {
-        throw new Error("value is not a valid HTTP-date or non-negative decimal integer");
-    }
-    return date.getTime() - Date.now();
-}
-
-/**
  * Retries a network operation run in a callback.
  * @param maxAttempts - maximum attempts to try
  * @param callback - callback that returns a promise of the network operation. If rejected with ConnectionError, it will be retried by calling the callback again.
